requirements.txt
opencv-python
python-chess
pyautogui
mss
numpy
pyttsx3
scikit-image
pygrabber
mac-say
berserk